Moved the batchnorm TPP headers to a separate folder op_lib_tpp
